* check the first item and dispatch to different loops

Signed-off-by: eat4toast <shihaohang@outlook.com>