@eat4toast 2022-06-23
## 0.  #1457 
This pr focus on solve meaningless branch in lib/executor/instantiate/function.cpp 
## 1. time improvement:

before:
```
$ time wasmedge_origin --dir .:. ./ruby test_base64.rb 
`RubyGems' were not loaded.
`error_highlight' was not loaded.
`did_you_mean' was not loaded.

real	0m13.929s
user	0m13.836s
sys	0m0.089s
```
after:
```
$ time wasmedge_new --dir .:. ./ruby test_base64.rb 
`RubyGems' were not loaded.
`error_highlight' was not loaded.
`did_you_mean' was not loaded.

real	0m13.401s
user	0m13.337s
sys	0m0.064s

```
## 2. branch miss improvement:

before:
```
$ sudo perf stat wasmedge_origin  --dir .:. ./ruby  test_base64.rb 
`RubyGems' were not loaded.
`error_highlight' was not loaded.
`did_you_mean' was not loaded.

 Performance counter stats for 'wasmedge_origin --dir .:. ./ruby test_base64.rb':

         13,770.29 msec task-clock                #    1.000 CPUs utilized          
                51      context-switches          #    3.704 /sec                   
                 7      cpu-migrations            #    0.508 /sec                   
           112,153      page-faults               #    8.145 K/sec                  
    64,530,033,135      cycles                    #    4.686 GHz                    
   176,930,521,638      instructions              #    2.74  insn per cycle         
    37,106,032,942      branches                  #    2.695 G/sec                  
       314,094,732      branch-misses             #    0.85% of all branches        
   322,621,373,105      slots                     #   23.429 G/sec                  
   156,833,268,341      topdown-retiring          #     47.5% retiring              
    61,941,467,307      topdown-bad-spec          #     18.7% bad speculation       
   102,396,103,994      topdown-fe-bound          #     31.0% frontend bound        
     9,202,768,311      topdown-be-bound          #      2.8% backend bound         

      13.771129138 seconds time elapsed

      13.670626000 seconds user
       0.100019000 seconds sys

```
after:
```
$ sudo perf stat wasmedge_new  --dir .:. ./ruby  test_base64.rb 
`RubyGems' were not loaded.
`error_highlight' was not loaded.
`did_you_mean' was not loaded.

 Performance counter stats for 'wasmedge_new --dir .:. ./ruby test_base64.rb':

         13,308.82 msec task-clock                #    1.000 CPUs utilized          
                57      context-switches          #    4.283 /sec                   
                 3      cpu-migrations            #    0.225 /sec                   
           112,151      page-faults               #    8.427 K/sec                  
    62,355,375,553      cycles                    #    4.685 GHz                    
   176,930,180,287      instructions              #    2.84  insn per cycle         
    37,105,976,298      branches                  #    2.788 G/sec                  
       329,381,495      branch-misses             #    0.89% of all branches        
   311,762,943,460      slots                     #   23.425 G/sec                  
   155,247,429,442      topdown-retiring          #     48.5% retiring              
    69,627,067,167      topdown-bad-spec          #     21.7% bad speculation       
    84,291,157,507      topdown-fe-bound          #     26.3% frontend bound        
    11,080,269,670      topdown-be-bound          #      3.5% backend bound         

      13.309779632 seconds time elapsed

      13.225129000 seconds user
       0.084007000 seconds sys

```

Since the whole branches is not same, so the branch-misses seems not improve well. But, from the #1457 and [preview](https://github.com/eat4toast/fix_addfunc_test_result/blob/478096a486c3ee9859985dda46db310358e06b0f/0621_test/log#L5),  running ruby container takes extra 6648 times, so it indeed need fix there.